Refactor: Apply ts to BrandFilter
