--- src/components/BrandFilter/index.js
@@ -1,3 +1,0 @@
-import BrandFilter from "./BrandFilter";
-
-export default BrandFilter;
