+++ src/components/BrandFilter/index.ts
@@ -1,0 +1,3 @@
+import BrandFilter from './BrandFilter';
+
+export default BrandFilter;
